requirements.txt
# Flask and Web Framework
Flask==3.0.0
flask-cors==4.0.0
flask-sock==0.7.0
gunicorn==21.2.0

# Data and Finance
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3

# Machine Learning and AI
scikit-learn==1.4.0
prophet==1.1.5
tensorflow==2.15.0
xgboost==2.0.3

# News and Sentiment Analysis
finnhub-python==2.4.19
textblob==0.17.1
newspaper3k==0.2.8

# API and HTTP
requests==2.31.0
python-dotenv==1.0.0

# Database (optional for caching)
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3